Replaced my work-around with an isolate and calls to Dart_PostCObject_DL to pass data back to the main thread.

+++ flutter/smartlock_idl_plugin/src/smartlock_idl_plugin.h
@@ -55,10 +55,8 @@ typedef struct {
 
   // In order to call back into Dart from another thread, it
   // must be done through a mechanism that can be forced through
-  // the main thread.  We should be able to use Dart_PostCObject_DL().
-  // But, that was causing a segmentation fault.  As a last resort,
-  // I fell back to sending data through a loopback socket.
-  short send_port;
+  // the main thread.  We will be using Dart_PostCObject_DL().
+  int64_t send_port;
 } OpenDdsBridgeConfig;
 
 FFI_PLUGIN_EXPORT OpenDdsBridge* createOpenDdsBridge();
